Convert preset DB Controller to preset controller - Move presetList into PresetsController from chordSelect screen - It is now more of state management of preset rather than just a database.

+++ lib/controllers/presets_controller.dart
@@ -5,9 +5,16 @@ import 'package:get/get.dart';
 import '../models/chord.dart';
 import '../models/preset.dart';
 
-class PresetDbController extends GetxController {
+/* 
+ * List of presets
+ * 
+ * Database opens when Get.put(PresetsController()) is called.
+ * Database closes when the page which calls Get.put(PresetsController()) destroys.
+ * 
+ * Calling open() or close() outside the class is unnecessary.
+ */
+class PresetsController extends GetxController {
   bool _isOpen = false;
-  bool _isInit = false;
   final String _tableName = 'Presets';
   final String _dbName = 'c2b_jaeiklee_chord_presets.db';
   final String _schema = '''(id INTEGER PRIMARY KEY AUTOINCREMENT,
@@ -15,53 +22,15 @@ class PresetDbController extends GetxController {
                              chords TEXT)''';
   late final String _path;
   late Database _db;
+  List<Preset> _presetList = [];
 
-  bool isOpen() {
-    return _isOpen;
-  }
-
-  Future<void> init() async {
-    await _setPath();
-    await _open();
-    _isInit = true;
-  }
-
-  Future<void> _setPath() async {
-    if (_isInit) return;
-    var path = await getDatabasesPath();
-    _path = join(path, _dbName);
-  }
-
-  Future<void> _open() async {
-    if (_isOpen) return;
-
-    _db = await openDatabase(
-      _path,
-      version: 1,
-      onCreate: (Database db, int version) async {
-        await db.execute('CREATE TABLE $_tableName $_schema');
-      },
-    );
-    _isOpen = true;
-  }
-
-  // Out of the class, `closeDb()` should be called nowhere
-  // but only in the top node of the widget tree
-  // since PresetDatabase is singleton.
-  Future<void> closeDb() async {
-    if (!_isOpen) return;
-
-    await _db.close();
-    _isOpen = false;
-  }
-
-  Future<void> cleanUpDb() async {
-    await closeDb();
-    await deleteDatabase(_path);
-    await _open();
+  List<Preset> get presetList {
+    return _presetList;
   }
 
   Future<void> saveAsPreset(String presetName, List<Chord> chords) async {
+    if (!_isOpen) return;
+
     final encodedPreset = _encode(chords);
     await _db.transaction((txn) async {
       var id = await txn.rawInsert(
@@ -69,25 +38,32 @@ class PresetDbController extends GetxController {
         [presetName, encodedPreset],
       );
     });
+    _fetchPresetList();
   }
 
   Future<void> deletePreset(int id) async {
+    if (!_isOpen) return;
+
     await _db.delete(
       _tableName,
       where: 'id = ?',
       whereArgs: [id],
     );
+    _fetchPresetList();
   }
 
-  Future<List<Preset>> getPresetList() async {
-    final presetList = await _db.rawQuery('SELECT * FROM $_tableName');
+  Future<void> _fetchPresetList() async {
+    if (!_isOpen) return;
 
-    return List.generate(presetList.length, (index) {
-      var id = presetList[index]['id'] as int;
-      var name = presetList[index]['name'] as String;
-      var chordList = _decode(presetList[index]['chords'] as String);
+    final result = await _db.rawQuery('SELECT * FROM $_tableName');
+
+    _presetList = List.generate(result.length, (index) {
+      var id = result[index]['id'] as int;
+      var name = result[index]['name'] as String;
+      var chordList = _decode(result[index]['chords'] as String);
       return Preset(id: id, name: name, chordList: chordList);
     });
+    update();
   }
 
   String _encode(List<Chord> chords) {
@@ -111,9 +87,49 @@ class PresetDbController extends GetxController {
     });
   }
 
+  Future<void> _init() async {
+    var path = await getDatabasesPath();
+    _path = join(path, _dbName);
+    await _open();
+    _fetchPresetList();
+  }
+
+  Future<void> _open() async {
+    if (_isOpen) return;
+
+    _db = await openDatabase(
+      _path,
+      version: 1,
+      onCreate: (Database db, int version) async {
+        await db.execute('CREATE TABLE $_tableName $_schema');
+      },
+    );
+    _isOpen = true;
+  }
+
+  Future<void> _close() async {
+    if (!_isOpen) return;
+
+    await _db.close();
+    _isOpen = false;
+  }
+
+  Future<void> clear() async {
+    await _close();
+    await deleteDatabase(_path);
+    await _open();
+    _fetchPresetList();
+  }
+
   @override
   void onInit() {
-    print('onInit preset db');
+    _init();
     super.onInit();
+  }
+
+  @override
+  void onClose() {
+    _close();
+    super.onClose();
   }
 }
